Highlighted that test data is a Splunk addon.
I spent a lot of time trying to figure out where the test data was and
how to get it into Splunk.  It was a total newbie mistake, and I felt
a bit abashed when I figured it out, but it seems like a mistake that
others could make as well, so I tried to clarify that a bit.